디렉토리 구조 변경
도메인별 디렉토리 추가
