--- src/main/java/codesquad/springcafe/article/DTO/ArticlePostReq.java
@@ -1,4 +1,0 @@
-package codesquad.springcafe.article.DTO;
-
-public record ArticlePostReq (String title, String content){
-}
