--- src/main/java/codesquad/springcafe/article/repository/ArticleConsts.java
@@ -1,5 +1,0 @@
-package codesquad.springcafe.article.repository;
-
-public abstract class ArticleConsts {
-    public static final int DEFAULT_POINT = 0;
-}
