--- src/main/java/codesquad/springcafe/article/repository/MemoryArticleRepository.java
@@ -1,76 +1,0 @@
-package codesquad.springcafe.article.repository;
-
-import codesquad.springcafe.article.Article;
-import codesquad.springcafe.article.DTO.ArticlePostReq;
-import codesquad.springcafe.user.DTO.SimpleUserInfo;
-import org.springframework.stereotype.Repository;
-
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import static codesquad.springcafe.article.repository.ArticleConsts.DEFAULT_POINT;
-
-@Repository
-public class MemoryArticleRepository implements ArticleRepository {
-
-    private final Map<Integer, Article> articles = new ConcurrentHashMap<>();
-    private int nowIndex = 1;
-
-    public MemoryArticleRepository() {
-        add(new ArticlePostReq(" 테스트용 제목", "테스트용 내용"), new SimpleUserInfo("테스터", "testerId"));
-    }
-
-    @Override
-    public void add(ArticlePostReq articlePostReq, SimpleUserInfo simpleUserInfo) throws IllegalArgumentException {
-        Article article = new Article(
-                nowIndex,
-                Timestamp.valueOf(LocalDateTime.now()),
-                simpleUserInfo,
-                articlePostReq.title(),
-                articlePostReq.content(),
-                DEFAULT_POINT);
-
-        articles.put(nowIndex, article);
-        nowIndex++;
-    }
-
-    @Override
-    public Article findById(int id) {
-        return articles.get(id);
-    }
-
-    @Override
-    public List<Article> findAll() {
-        return new ArrayList<>(articles.values());
-    }
-
-    @Override
-    public List<Article> findByUserId(String userId) {
-        return articles.values().stream().filter(a -> a.getAuthor().id().equals(userId)).toList();
-    }
-
-    @Override
-    public void addPoint(int articleId) {
-        articles.get(articleId).addPoint();
-    }
-
-    @Override
-    public void update(int id, ArticlePostReq articlePostReq) {
-        SimpleUserInfo author = articles.remove(id).getAuthor();
-        add(articlePostReq, author);
-    }
-
-    @Override
-    public void delete(int id) {
-        articles.remove(id);
-    }
-
-    @Override
-    public void deleteAll() {
-        articles.clear();
-    }
-}
