--- src/main/java/codesquad/springcafe/comment/DTO/CommentPostReq.java
@@ -1,4 +1,0 @@
-package codesquad.springcafe.comment.DTO;
-
-public record CommentPostReq(int articleId, String content) {
-}
