--- src/main/java/codesquad/springcafe/comment/repository/CommentRepository.java
@@ -1,22 +1,0 @@
-package codesquad.springcafe.comment.repository;
-
-import codesquad.springcafe.comment.DTO.Comment;
-import codesquad.springcafe.comment.DTO.CommentPostReq;
-import codesquad.springcafe.user.DTO.SimpleUserInfo;
-
-import java.util.List;
-
-public interface CommentRepository {
-
-
-    void add(CommentPostReq commentPostReq, SimpleUserInfo simpleUserInfo) throws IllegalArgumentException;
-
-    List<Comment> findByArticleId(int id);
-
-    List<Comment> findByUserId(String id);
-    void modify(int id, CommentPostReq commentPostReq);
-    void delete(int id);
-
-    Comment findById(int id);
-}
-
