+++ src/main/java/codesquad/springcafe/domain/article/DTO/ArticlePostReq.java
@@ -1,0 +1,4 @@
+package codesquad.springcafe.domain.article.DTO;
+
+public record ArticlePostReq (String title, String content){
+}
