+++ src/main/java/codesquad/springcafe/domain/article/MainController.java
@@ -1,0 +1,26 @@
+package codesquad.springcafe.domain.article;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/")
+public class MainController {
+
+    private final ArticleService articleService;
+
+    @Autowired
+    public MainController(ArticleService articleService) {
+        this.articleService = articleService;
+    }
+
+
+    @GetMapping("")
+    public String index(Model model){
+        model.addAttribute("articles", articleService.findAll());
+        return "index";
+    }
+}
