+++ src/main/java/codesquad/springcafe/domain/article/repository/ArticleConsts.java
@@ -1,0 +1,5 @@
+package codesquad.springcafe.domain.article.repository;
+
+public abstract class ArticleConsts {
+    public static final int DEFAULT_POINT = 0;
+}
