+++ src/main/java/codesquad/springcafe/domain/comment/DTO/CommentPostReq.java
@@ -1,0 +1,4 @@
+package codesquad.springcafe.domain.comment.DTO;
+
+public record CommentPostReq(int articleId, String content) {
+}
