+++ src/main/java/codesquad/springcafe/domain/comment/repository/CommentRepository.java
@@ -1,0 +1,22 @@
+package codesquad.springcafe.domain.comment.repository;
+
+import codesquad.springcafe.domain.comment.DTO.Comment;
+import codesquad.springcafe.domain.comment.DTO.CommentPostReq;
+import codesquad.springcafe.domain.user.DTO.SimpleUserInfo;
+
+import java.util.List;
+
+public interface CommentRepository {
+
+
+    void add(CommentPostReq commentPostReq, SimpleUserInfo simpleUserInfo) throws IllegalArgumentException;
+
+    List<Comment> findByArticleId(int id);
+
+    List<Comment> findByUserId(String id);
+    void modify(int id, CommentPostReq commentPostReq);
+    void delete(int id);
+
+    Comment findById(int id);
+}
+
