+++ src/main/java/codesquad/springcafe/domain/user/DTO/Alert.java
@@ -1,0 +1,4 @@
+package codesquad.springcafe.domain.user.DTO;
+
+public record Alert(String message) {
+}
