+++ src/main/java/codesquad/springcafe/domain/user/DTO/SimpleUserInfo.java
@@ -1,0 +1,4 @@
+package codesquad.springcafe.domain.user.DTO;
+
+public record SimpleUserInfo(String id , String name) {
+}
