+++ src/main/java/codesquad/springcafe/domain/user/DTO/UserListRes.java
@@ -1,0 +1,4 @@
+package codesquad.springcafe.domain.user.DTO;
+
+public record UserListRes(int index, String id, String name, String email) {
+}
