+++ src/main/java/codesquad/springcafe/domain/user/User.java
@@ -1,0 +1,45 @@
+package codesquad.springcafe.domain.user;
+
+public class User {
+    private String userId;
+    private String password;
+    private String name;
+    private String email;
+
+    public User(String userId, String password, String name, String email) {
+        this.userId = userId;
+        this.password = password;
+        this.name = name;
+        this.email = email;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public boolean isCorrectPassword(String password) {
+        return this.password.equals(password);
+    }
+
+    @Override
+    public String toString() {
+        return "User{" +
+                "id='" + userId + '\'' +
+                ", password='" + password + '\'' +
+                ", name='" + name + '\'' +
+                ", email='" + email + '\'' +
+                '}';
+    }
+}
