+++ src/main/java/codesquad/springcafe/domain/user/interceptor/NavbarSetInterceptor.java
@@ -1,0 +1,19 @@
+package codesquad.springcafe.domain.user.interceptor;
+
+import codesquad.springcafe.domain.user.DTO.SimpleUserInfo;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.ModelAndView;
+
+public class NavbarSetInterceptor implements HandlerInterceptor {
+
+    @Override
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) {
+        SimpleUserInfo loginUser = (SimpleUserInfo) request.getSession().getAttribute("loginUser");
+        if (modelAndView != null && loginUser != null) {
+            modelAndView.addObject("navbarUser", loginUser);
+            modelAndView.addObject("user_section", true);
+        }
+    }
+}
