--- src/main/java/codesquad/springcafe/user/DTO/Alert.java
@@ -1,4 +1,0 @@
-package codesquad.springcafe.user.DTO;
-
-public record Alert(String message) {
-}
