--- src/main/java/codesquad/springcafe/user/DTO/SimpleUserInfo.java
@@ -1,4 +1,0 @@
-package codesquad.springcafe.user.DTO;
-
-public record SimpleUserInfo(String id , String name) {
-}
