--- src/main/java/codesquad/springcafe/user/DTO/UserListRes.java
@@ -1,4 +1,0 @@
-package codesquad.springcafe.user.DTO;
-
-public record UserListRes(int index, String id, String name, String email) {
-}
